+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -59,9 +59,9 @@ class MenuServiceTest {
     @BeforeEach
     void setUp() {
         메뉴그룹_한식 = createMenuGroup(1L, "한식메뉴");
-        김치찌개 = createProduct(1L, "김치찌개", BigDecimal.valueOf(8000L));
-        공기밥 = createProduct(2L, "공기밥", BigDecimal.valueOf(1000L));
-        메뉴_김치찌개세트 = createMenu(1L, "김치찌개세트", BigDecimal.valueOf(15000L), 메뉴그룹_한식);
+        김치찌개 = createProduct(1L, "김치찌개", 8000);
+        공기밥 = createProduct(2L, "공기밥", 1000);
+        메뉴_김치찌개세트 = createMenu(1L, "김치찌개세트", 15000, 메뉴그룹_한식);
 
         김치찌개세트_김치찌개 = createMenuProduct(1L, 메뉴_김치찌개세트, 김치찌개, 2);
         김치찌개세트_공기밥 = createMenuProduct(1L, 메뉴_김치찌개세트, 공기밥, 2);
@@ -79,7 +79,7 @@ class MenuServiceTest {
         given(menuRepository.save(any(Menu.class))).willReturn(메뉴_김치찌개세트);
         MenuRequest 메뉴_김치찌개세트_request = MenuRequest.of(
                 메뉴_김치찌개세트.getName(),
-                메뉴_김치찌개세트.getPrice().intValue(),
+                메뉴_김치찌개세트.getPrice(),
                 메뉴_김치찌개세트.getMenuGroup().getId(),
                 메뉴_김치찌개세트.getMenuProducts().stream().
                         map(menuProduct -> MenuProductRequest.of(menuProduct.getProduct().getId(), (int) menuProduct.getQuantity()))
@@ -141,7 +141,7 @@ class MenuServiceTest {
         //when
         MenuRequest invalidMenu = MenuRequest.of(
                 메뉴_김치찌개세트.getName(),
-                메뉴_김치찌개세트.getPrice().intValue(),
+                메뉴_김치찌개세트.getPrice(),
                 메뉴_김치찌개세트.getMenuGroup().getId(),
                 메뉴_김치찌개세트.getMenuProducts().stream().
                         map(menuProduct -> MenuProductRequest.of(menuProduct.getProduct().getId(), (int) menuProduct.getQuantity()))
@@ -160,7 +160,7 @@ class MenuServiceTest {
         given(productRepository.findById(김치찌개.getId())).willReturn(Optional.ofNullable(null));
         MenuRequest invalidMenu = MenuRequest.of(
                 메뉴_김치찌개세트.getName(),
-                메뉴_김치찌개세트.getPrice().intValue(),
+                메뉴_김치찌개세트.getPrice(),
                 메뉴_김치찌개세트.getMenuGroup().getId(),
                 메뉴_김치찌개세트.getMenuProducts().stream().
                         map(menuProduct -> MenuProductRequest.of(menuProduct.getProduct().getId(), (int) menuProduct.getQuantity()))
